refactor account canister methods
