+++ src/canisters/accounts/src/api/account/mod.rs
@@ -1,2 +1,3 @@
 pub mod aegis_account;
 pub mod aegis_account_info;
+
