+++ src/canisters/accounts/src/api/interfaces/ledger.rs
@@ -3,6 +3,8 @@ use ic_ledger_utils::types::icrc_types::IcrcTransferResult;
 use minter_utils::services::ckbtc::{RetrieveBtcError, RetrieveBtcOk};
 use serde::Deserialize;
 
+use super::constants::CanisterName;
+
 #[derive(CandidType, Deserialize, PartialEq, Eq, Hash)]
 pub enum ICRCLedgerType {
     ICP,
@@ -23,4 +25,20 @@ pub enum RetrieveBtcResult {
     RetrieveBtcOk(RetrieveBtcOk),
     RetrieveBtcError(RetrieveBtcError),
     RetrieveBtcString(String),
+}
+
+#[derive(CandidType, Deserialize)]
+pub enum Ledger {
+    ICRC(CanisterName),
+    ETH,
+    BTC,
+}
+
+impl Ledger {
+    pub fn get_canister_name(&self) -> Option<CanisterName> {
+        match self {
+            Ledger::ICRC(canister_name) => Some(canister_name.clone()),
+            _ => None,
+        }
+    }
 }
