+++ src/canisters/accounts/src/api/wallet/icrc/mod.rs
@@ -1,0 +1,2 @@
+pub mod icrc_ledger;
+pub mod minter;
