+++ src/canisters/accounts/src/queries/account/mod.rs
@@ -1,0 +1,1 @@
+pub mod account;
