+++ src/canisters/accounts/src/queries/mod.rs
@@ -2,3 +2,4 @@ pub mod uitls;
 pub mod get_metrics;
 pub mod account;
 pub mod constants;
+pub mod wallet;
