+++ src/canisters/accounts/src/queries/wallet/icrc/mod.rs
@@ -1,0 +1,1 @@
+pub mod icrc_ledger;
