+++ src/canisters/accounts/src/queries/wallet/mod.rs
@@ -1,0 +1,1 @@
+pub mod icrc;
