+++ src/canisters/accounts/src/updates/account/mod.rs
@@ -1,0 +1,1 @@
+pub mod account;
