+++ src/canisters/accounts/src/updates/wallet/icrc/mod.rs
@@ -1,0 +1,2 @@
+pub mod icrc_ledger;
+pub mod minter;
