+++ src/canisters/accounts/src/updates/wallet/mod.rs
@@ -1,0 +1,1 @@
+pub mod icrc;
